fix: Update OmniDrive velocity handling and improve Robot sensor transmission

+++ 2026/Firmware/MainBoard/MainBoard_V26_1_1/src/unit/omni_drive.h
@@ -15,7 +15,7 @@ typedef struct {
 
 typedef struct {
   Serial *serials[4];
-  int16_t vel_wheel_angular[4];
+  float vel_wheel_angular[4];
   uint8_t emg;
   uint8_t ready;
 } OmniDrive;
